refactor(activerecord): route merge_joins folds through joins!/left_outer_joins!
Inline foldMergeJoins/foldMergeOuterJoins back into Merger#mergeJoins /
leftOuterJoinsBang, mirroring merger.rb:117-152. Drops the duplicate
joinsUnionEq structural-union loop in favour of joins!'s own dedup.
