+++ scripts/api-compare/call-mismatches-wide-exclude/activerecord/relation/merger.json
@@ -17,34 +17,6 @@
     "package": "activerecord",
     "tsFile": "relation/merger.ts",
     "rubyName": "merge_joins",
-    "call": "construct_join_dependency",
-    "reason": "Wide call-set flag: Merger#mergeJoins/#mergeOuterJoins are thin wrappers delegating to the shared foldMergeJoins/foldMergeOuterJoins helper (merge-joins.ts); the Rails call is satisfied inside the helper. PR #4676."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_joins",
-    "call": "joins!",
-    "reason": "Per-entry verified (RFC 0083 step 3): Rails merger.rb:117-134/136-152 ends the cross-model branch with `relation.joins!(join_dependency, *others)` / `relation.left_outer_joins!(...)`; trails merger.ts#mergeJoins/#mergeOuterJoins are thin wrappers over foldMergeJoins/foldMergeOuterJoins (relation/merge-joins.ts), which append to `_joinsValues` / `_leftOuterJoinsValues` with the same structural union `joins!` performs. The wide gate does not follow into the helper, so the entry stays until the helper split itself is converged (`converge-merge-joins-helper-onto-joins-bang`)."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_joins",
-    "call": "joins_values",
-    "reason": "Wide call-set flag: Merger#mergeJoins/#mergeOuterJoins are thin wrappers delegating to the shared foldMergeJoins/foldMergeOuterJoins helper (merge-joins.ts); the Rails call is satisfied inside the helper. PR #4676."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_joins",
-    "call": "model",
-    "reason": "Per-entry verified (RFC 0072 cross-helper reads): Rails merger.rb:117-134 reads `other.model` / `relation.model` while folding joins; trails merger.ts#mergeJoins is a thin wrapper over `foldMergeJoins` (relation/merge-joins.ts), which performs the same-model comparison through `.model` (PR #5325). The folder is shared with mergeOuterJoins, so the split stays."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_joins",
     "call": "partition",
     "reason": "Baseline (RFC 0047): wide call-set flag seeded when the wide ratchet landed; bucket (b) equivalent or (c) noise pending per-cluster burndown review."
   },
@@ -89,34 +61,6 @@
     "rubyName": "merge_multi_values",
     "call": "reordering_value",
     "reason": "Baseline (RFC 0047): wide call-set flag seeded when the wide ratchet landed; bucket (b) equivalent or (c) noise pending per-cluster burndown review."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_outer_joins",
-    "call": "construct_join_dependency",
-    "reason": "Wide call-set flag: Merger#mergeJoins/#mergeOuterJoins are thin wrappers delegating to the shared foldMergeJoins/foldMergeOuterJoins helper (merge-joins.ts); the Rails call is satisfied inside the helper. PR #4676."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_outer_joins",
-    "call": "left_outer_joins!",
-    "reason": "Per-entry verified (RFC 0083 step 3): Rails merger.rb:117-134/136-152 ends the cross-model branch with `relation.joins!(join_dependency, *others)` / `relation.left_outer_joins!(...)`; trails merger.ts#mergeJoins/#mergeOuterJoins are thin wrappers over foldMergeJoins/foldMergeOuterJoins (relation/merge-joins.ts), which append to `_joinsValues` / `_leftOuterJoinsValues` with the same structural union `joins!` performs. The wide gate does not follow into the helper, so the entry stays until the helper split itself is converged (`converge-merge-joins-helper-onto-joins-bang`)."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_outer_joins",
-    "call": "left_outer_joins_values",
-    "reason": "Wide call-set flag: Merger#mergeJoins/#mergeOuterJoins are thin wrappers delegating to the shared foldMergeJoins/foldMergeOuterJoins helper (merge-joins.ts); the Rails call is satisfied inside the helper. PR #4676."
-  },
-  {
-    "package": "activerecord",
-    "tsFile": "relation/merger.ts",
-    "rubyName": "merge_outer_joins",
-    "call": "model",
-    "reason": "Per-entry verified (RFC 0072 cross-helper reads): Rails merger.rb:136-152 reads `other.model` / `relation.model` while folding left outer joins; trails merger.ts#mergeOuterJoins is a thin wrapper over `foldMergeOuterJoins` (relation/merge-joins.ts), which performs the same-model comparison through `.model` (PR #5325). The folder is shared with mergeJoins, so the split stays."
   },
   {
     "package": "activerecord",
