+++ scripts/api-compare/call-mismatches-wide-unreviewed.json
@@ -1,3 +1,3 @@
 {
-  "max": 2845
+  "max": 2840
 }
